Fix for explanation gather issue (#7)
There was a problem with the gathering of explanations. Basically, because even an empty `RAIInsights` object will always have a file in its explanation directory, when we copied the one we had just computed into position, the `RAIInsights` object which we were building would have two GUID directories under `explainer`. Whether or not the explanation then loaded was down to which one happened to come up first - the 'empty' explanation or the computed one. Since the Gather component then saved the loaded explanation, the resultant output port might not have the one just computed.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>