Adjust output of the Constructor component (#28)
Change the output of the RAI Constructor component.

Previously, the output of the Constructor was the result of calling `RAIInsights.save()`. This meant that any special handling (especially for the supplied model) would only be possible within the Constructor; thereafter, the model would be pickled and passed along to the subsequent components for use.

This changeset makes the Constructor copy its inputs (the model information file and the two datasets) to its output, and add in a JSON serialisation of the inputs parameters (such as the list of categorical features). Each individual tool component can use this to create its own `RAIInsights` object, and then work as before. The `Gather` component also requires some small changes.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>