Single RAI Job for UX (#35)
This single `CommandJob` bypasses all the components to enable our UX scenario prior to a full component solution being available. It serves as a template for embedding into the UX.

This change copies code from the components. Needless to say, this is not ideal. It also hardcodes links to particular versions of datasets, and so requires manual updating before use.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>