Pipeline fixes for v2.1.2 (#49)
There is an issue making copies of the incoming data when the datasets are present via `ro_mount`. The files are copied via `shutil.copytree()` and something inside that code is getting confused. Since the current approach is going to be retired once `MLTable` comes online, just switch to `download` mode for now.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>