Upper bound on SciKit-Learn to fix builds (#81)
The latest release of SciKit-Learn is causing trouble in the causal component. See the [PR in `responsible-ai-toolbox`](https://github.com/microsoft/responsible-ai-toolbox/pull/1429) for details. Until a new version of `responsibleai` is available, we have to apply the upper bound in this repo too.

Also fix a path issue in one of the notebook tests.

Signed-off-by: Richard Edgar <riedgar@microsoft.com>